refactor(frontend): optimize ECharts imports
