+++ frontend/app/src/premium/setup-premium.ts
@@ -1,12 +1,36 @@
 import * as VueUse from '@vueuse/core';
 import * as VueUseShared from '@vueuse/shared';
 import * as BigNumber from 'bignumber.js';
-import * as ECharts from 'echarts';
+import { BarChart, LineChart, PieChart } from 'echarts/charts';
+import {
+  BrushComponent,
+  DataZoomComponent,
+  GridComponent,
+  LegendComponent,
+  ToolboxComponent,
+  TooltipComponent,
+} from 'echarts/components';
+import * as EChartsCore from 'echarts/core';
+import { CanvasRenderer } from 'echarts/renderers';
 import * as Vue from 'vue';
 import VChart from 'vue-echarts';
 import * as VueRouter from 'vue-router';
 import * as zod from 'zod/v4';
 import { app } from '@/main';
+
+// Register only the needed echarts components for premium
+EChartsCore.use([
+  CanvasRenderer,
+  LineChart,
+  PieChart,
+  BarChart,
+  TooltipComponent,
+  GridComponent,
+  DataZoomComponent,
+  LegendComponent,
+  BrushComponent,
+  ToolboxComponent,
+]);
 
 export async function setupPremium(): Promise<void> {
   /**
@@ -17,7 +41,7 @@ export async function setupPremium(): Promise<void> {
     window.VueUse = VueUse;
     window.VueUseShared = VueUseShared;
     window.VueEcharts = VChart;
-    window.ECharts = ECharts;
+    window.ECharts = EChartsCore;
     window.zod = zod;
     window.bn = BigNumber;
     window.VueRouter = VueRouter;
